+++ examples/plot_chord.py
@@ -5,6 +5,7 @@ Chord Diagram
 import numpy as np
 
 import msmexplorer as msme
+from msmexplorer.utils import make_colormap
 
 
 # Create a random square matrix
@@ -13,5 +14,8 @@ n, p = 12, 12
 d = rs.normal(0, 2, (n, p))
 d += np.log(np.arange(1, p + 1)) * -5 + 10
 
+# Make a colormap
+cmap = make_colormap(['rawdenim', 'lightgray', 'pomegranate'])
+
 # Plot Chord Diagram
-msme.plot_chord(d, threshold=.2)
+msme.plot_chord(d, cmap=cmap, threshold=.2)