+++ examples/plot_free_energy_2d.py
@@ -2,7 +2,6 @@
 Free Energy Plot (Bivariate)
 ============================
 """
-from msmbuilder.example_datasets import FsPeptide
 from msmbuilder.featurizer import DihedralFeaturizer
 from msmbuilder.decomposition import tICA
 from msmbuilder.cluster import MiniBatchKMeans
@@ -11,11 +10,12 @@ from msmbuilder.msm import MarkovStateModel
 import numpy as np
 
 import msmexplorer as msme
+from msmexplorer.example_datasets import FsPeptide
 
-np.random.seed(42)
+rs = np.random.RandomState(42)
 
 # Load Fs Peptide Data
-trajs = FsPeptide().get().trajectories[:10]
+trajs = FsPeptide().get().trajectories
 
 # Extract Backbone Dihedrals
 featurizer = DihedralFeaturizer(types=['phi', 'psi'])
@@ -26,14 +26,15 @@ tica_model = tICA(lag_time=2, n_components=2)
 tica_trajs = tica_model.fit_transform(diheds)
 
 # Perform Clustering
-clusterer = MiniBatchKMeans(n_clusters=12)
+clusterer = MiniBatchKMeans(n_clusters=12, random_state=rs)
 clustered_trajs = clusterer.fit_transform(tica_trajs)
 
 # Construct MSM
-msm = MarkovStateModel(lag_time=2, n_timescales=5)
+msm = MarkovStateModel(lag_time=2)
 assignments = msm.fit_transform(clustered_trajs)
 
 # Plot Free Energy
 data = np.concatenate(tica_trajs, axis=0)
 pi_0 = msm.populations_[np.concatenate(assignments, axis=0)]
-msme.plot_free_energy(data, obs=(0, 1), n_samples=100000, pi=pi_0)
+msme.plot_free_energy(data, obs=(0, 1), n_samples=100000, pi=pi_0,
+                      random_state=rs)