+++ examples/plot_histogram.py
@@ -2,15 +2,13 @@
 Histogram Plot
 ==============
 """
-from msmbuilder.example_datasets import FsPeptide
 from msmbuilder.featurizer import DihedralFeaturizer
 from msmbuilder.decomposition import tICA
 
 import numpy as np
 
 import msmexplorer as msme
-
-np.random.seed(42)
+from msmexplorer.example_datasets import FsPeptide
 
 # Load Fs Peptide Data
 trajs = FsPeptide().get().trajectories
@@ -25,4 +23,6 @@ tica_trajs = tica_model.fit_transform(diheds)
 
 # Plot Histogram
 data = np.concatenate(tica_trajs, axis=0)
-msme.plot_histogram(data)
+msme.plot_histogram(data, color='oxblood', quantiles=(0.5,),
+                    labels=['$tIC1$', '$tIC2$', '$tIC3$', '$tIC4$'],
+                    show_titles=True)