+++ examples/plot_tpaths.py
@@ -1,8 +1,7 @@
 """
-MSM Network Plot
-================
+Transition Pathway Plot
+=======================
 """
-from msmbuilder.example_datasets import FsPeptide
 from msmbuilder.featurizer import DihedralFeaturizer
 from msmbuilder.decomposition import tICA
 from msmbuilder.cluster import MiniBatchKMeans
@@ -11,9 +10,9 @@ from msmbuilder.msm import MarkovStateModel
 import numpy as np
 
 import msmexplorer as msme
-from msmexplorer.palettes import msme_rgb
+from msmexplorer.example_datasets import FsPeptide
 
-np.random.seed(42)
+rs = np.random.RandomState(42)
 
 # Load Fs Peptide Data
 trajs = FsPeptide().get().trajectories
@@ -23,18 +22,18 @@ featurizer = DihedralFeaturizer(types=['phi', 'psi'])
 diheds = featurizer.fit_transform(trajs)
 
 # Perform Dimensionality Reduction
-tica_model = tICA(lag_time=2, n_components=4)
+tica_model = tICA(lag_time=2, n_components=2)
 tica_trajs = tica_model.fit_transform(diheds)
 
 # Perform Clustering
-clusterer = MiniBatchKMeans(n_clusters=100)
+clusterer = MiniBatchKMeans(n_clusters=100, random_state=rs)
 clustered_trajs = clusterer.fit_transform(tica_trajs)
 
 # Construct MSM
-msm = MarkovStateModel(lag_time=2, n_timescales=5)
+msm = MarkovStateModel(lag_time=2)
 msm.fit(clustered_trajs)
 
 # Plot TPT
-msme.plot_tpaths(msm, sources=[99], sinks=[0],
-                 node_color=msme_rgb['beryl'],
-                 edge_color=msme_rgb['dark slate grey'])
+pos = dict(zip(range(clusterer.n_clusters), clusterer.cluster_centers_))
+msme.plot_tpaths(msm, [99], [0], pos=pos, node_color='beryl',
+                 edge_color='carbon')