+++ msmexplorer/plots/msm.py
@@ -1,14 +1,15 @@
 import numpy as np
 import networkx as nx
-import seaborn.apionly as sns
+import seaborn as sns
 from matplotlib import pyplot as pp
 
-from ..utils import extract_palette
+from ..utils import msme_colors
 from ..palettes import msme_rgb
 
 __all__ = ['plot_pop_resids', 'plot_msm_network', 'plot_timescales']
 
 
+@msme_colors
 def plot_pop_resids(msm, **kwargs):
     """
     Plot residuals between MSM populations and raw counts.
@@ -40,8 +41,10 @@ def plot_pop_resids(msm, **kwargs):
     return ax
 
 
-def plot_msm_network(msm, pos=None, node_color='c', node_size=300,
-                     edge_color='k', ax=None, with_labels=True, **kwargs):
+@msme_colors
+def plot_msm_network(msm, pos=None, node_size=None, node_color='pomegranate',
+                     edge_color='carbon', alpha=.7, ax=None, with_labels=True,
+                     **kwargs):
     """
     Plot MSM network diagram.
 
@@ -53,14 +56,16 @@ def plot_msm_network(msm, pos=None, node_color='c', node_size=300,
         Node positions in dict format (e.g. {node_id : [x, y]})
     node_color : str or [r, g, b], optional
         networkx node colors
-    node_size : int, optional
+    node_size : int or list, optional
         networkx node size
-    node_size : str or [r, g, b], optional
+    node_color : str, optional
         networkx edge color
     ax : matplotlib axis, optional (default: None)
         Axis to plot on, otherwise uses current axis.
     with_labels : boolean, optional
         Whether or not to include node labels (default: True)
+    alpha : float, optional  (default: 0.7)
+        Opacity of nodes
     **kwargs : dict, optional
         Extra arguments to pass to networkx.draw_networkx
 
@@ -72,20 +77,36 @@ def plot_msm_network(msm, pos=None, node_color='c', node_size=300,
     """
     if hasattr(msm, 'all_populations_'):
         tmat = msm.all_transmats_.mean(0)
+        pop = msm.all_populations_.mean(0)
     elif hasattr(msm, 'populations_'):
         tmat = msm.transmat_
+        pop = msm.populations_
 
     graph = nx.Graph(tmat)
 
     if not ax:
         ax = pp.gca()
 
+    if pos is None:
+        pos = nx.spring_layout(graph)
+
+    if node_size is None:
+        node_size = 5000. * pop
+
+    if isinstance(node_size, (list, np.ndarray)):
+        node_size = [node_size[i] for i in graph.nodes()]
+
+    if isinstance(node_color, (list, np.ndarray)):
+        node_color = [node_color[i] for i in graph.nodes()]
+
     nx.draw_networkx(graph, pos=pos, node_color=node_color,
-                     edge_color=edge_color, ax=ax, **kwargs)
+                     node_size=node_size, edge_color=edge_color, ax=ax,
+                     with_labels=with_labels, alpha=alpha, **kwargs)
 
     return ax
 
 
+@msme_colors
 def plot_timescales(msm, n_timescales=None, error=None, sigma=2,
                     color_palette=None, xlabel=None, ylabel=None, ax=None):
     """
@@ -138,13 +159,11 @@ def plot_timescales(msm, n_timescales=None, error=None, sigma=2,
     if not ax:
         _, ax = pp.subplots(1, 1, figsize=(2, 8))
     if not color_palette:
-        colors = list(msme_rgb.values())
-    else:
-        colors = extract_palette(color_palette)
+        color_palette = list(msme_rgb.values())
 
     for i, item in enumerate(zip(timescales, error)):
         t, s = item
-        color = colors[i % len(colors)]
+        color = color_palette[i % len(color_palette)]
         ax.errorbar([0, 1], [t, t], c=color)
         if s:
             for j in range(1, sigma + 1):
@@ -154,7 +173,8 @@ def plot_timescales(msm, n_timescales=None, error=None, sigma=2,
 
     ax.xaxis.set_ticks([])
     if xlabel:
-        ax.xaxis.set_label_text(xlabel, size=18, labelpad=18)
+        ax.xaxis.set_label_text(xlabel, size=18)
+        ax.xaxis.labelpad = 18
     if ylabel:
         ax.yaxis.set_label_text(ylabel, size=18)
     ax.set_yscale('log')